+++ tests/test_formula_printer.py
@@ -12,7 +12,7 @@ from snuba_sdk.timeseries import Metric, Timeseries
 formula_tests = [
     pytest.param(
         Formula(
-            ArithmeticOperator.DIVIDE,
+            ArithmeticOperator.DIVIDE.value,
             [
                 Timeseries(
                     metric=Metric(
@@ -28,7 +28,7 @@ formula_tests = [
     ),
     pytest.param(
         Formula(
-            ArithmeticOperator.MULTIPLY,
+            ArithmeticOperator.MULTIPLY.value,
             [
                 Timeseries(
                     metric=Metric(
@@ -45,13 +45,14 @@ formula_tests = [
             ],
         ),
         "(sum(foo) * max(bar))",
+        id="test two aggregate terms",
     ),
     pytest.param(
         Formula(
-            ArithmeticOperator.DIVIDE,
+            ArithmeticOperator.DIVIDE.value,
             [
                 Formula(
-                    ArithmeticOperator.MULTIPLY,
+                    ArithmeticOperator.MULTIPLY.value,
                     [
                         Timeseries(
                             metric=Metric(
@@ -73,10 +74,11 @@ formula_tests = [
             ],
         ),
         "((sum(foo) * sum(bar)) / 1000.0)",
+        id="test nested terms",
     ),
     pytest.param(
         Formula(
-            ArithmeticOperator.DIVIDE,
+            ArithmeticOperator.DIVIDE.value,
             [
                 Timeseries(
                     metric=Metric(
@@ -94,10 +96,11 @@ formula_tests = [
             filters=[Condition(Column("tag"), Op.EQ, "tag_value")],
         ),
         '(sum(foo) / sum(bar)){tag:"tag_value"}',
+        id="test terms with filter",
     ),
     pytest.param(
         Formula(
-            ArithmeticOperator.DIVIDE,
+            ArithmeticOperator.DIVIDE.value,
             [
                 Timeseries(
                     metric=Metric(
@@ -116,10 +119,11 @@ formula_tests = [
             ],
         ),
         '(sum(foo){tag:"tag_value"} / sum(bar){tag:"tag_value"})',
+        id="test terms with filters",
     ),
     pytest.param(
         Formula(
-            ArithmeticOperator.DIVIDE,
+            ArithmeticOperator.DIVIDE.value,
             [
                 Timeseries(
                     metric=Metric(
@@ -138,10 +142,11 @@ formula_tests = [
             groupby=[Column("transaction")],
         ),
         '(sum(foo) / sum(bar)){tag:"tag_value"} by (transaction)',
+        id="test terms with filters and groupby",
     ),
     pytest.param(
         Formula(
-            ArithmeticOperator.DIVIDE,
+            ArithmeticOperator.DIVIDE.value,
             [
                 Timeseries(
                     metric=Metric(
@@ -160,10 +165,11 @@ formula_tests = [
             ],
         ),
         "(sum(foo) by (transaction) / sum(bar) by (transaction))",
+        id="test terms with filters and groupbys",
     ),
     pytest.param(
         Formula(
-            ArithmeticOperator.DIVIDE,
+            ArithmeticOperator.DIVIDE.value,
             [
                 Timeseries(
                     metric=Metric(
@@ -183,10 +189,11 @@ formula_tests = [
             filters=[Condition(Column("tag"), Op.EQ, "tag_value")],
         ),
         '(sum(foo) by (transaction) / sum(bar) by (transaction)){tag:"tag_value"}',
+        id="test terms with inner groupbys and outer filter",
     ),
     pytest.param(
         Formula(
-            ArithmeticOperator.DIVIDE,
+            ArithmeticOperator.DIVIDE.value,
             [
                 Timeseries(
                     metric=Metric(
@@ -207,34 +214,11 @@ formula_tests = [
             ],
         ),
         '(sum(foo){tag:"tag_value"} by (transaction) / sum(bar){tag:"tag_value"} by (transaction))',
+        id="test terms with inner groupbys and inner filter",
     ),
     pytest.param(
         Formula(
-            ArithmeticOperator.DIVIDE,
-            [
-                Timeseries(
-                    metric=Metric(
-                        public_name="foo", entity="generic_metrics_distributions"
-                    ),
-                    aggregate="sum",
-                    filters=[Condition(Column("tag"), Op.EQ, "tag_value")],
-                    groupby=[Column("transaction")],
-                ),
-                Timeseries(
-                    metric=Metric(
-                        public_name="bar", entity="generic_metrics_distributions"
-                    ),
-                    aggregate="sum",
-                    filters=[Condition(Column("tag"), Op.EQ, "tag_value")],
-                    groupby=[Column("transaction")],
-                ),
-            ],
-        ),
-        '(sum(foo){tag:"tag_value"} by (transaction) / sum(bar){tag:"tag_value"} by (transaction))',
-    ),
-    pytest.param(
-        Formula(
-            ArithmeticOperator.DIVIDE,
+            ArithmeticOperator.DIVIDE.value,
             [
                 Timeseries(
                     metric=Metric(
@@ -253,13 +237,14 @@ formula_tests = [
             groupby=[Column("transaction")],
         ),
         '(sum(foo) / sum(bar)){tag:"tag_value"} by (transaction)',
+        id="test terms with outer groupbys and outer filter",
     ),
     pytest.param(
         Formula(
-            operator=ArithmeticOperator.MULTIPLY,
+            function_name=ArithmeticOperator.MULTIPLY.value,
             parameters=[
                 Formula(
-                    ArithmeticOperator.DIVIDE,
+                    ArithmeticOperator.DIVIDE.value,
                     [
                         Timeseries(
                             metric=Metric(
@@ -292,13 +277,14 @@ formula_tests = [
             groupby=[Column("transaction")],
         ),
         '((sum(foo){tag2:"tag_value2" AND tag:"tag_value"} / sum(bar)){tag3:"tag_value3"} * sum(pop)) by (transaction)',
+        id="test complex nested terms",
     ),
     pytest.param(
         Formula(
-            operator=ArithmeticOperator.MULTIPLY,
+            function_name=ArithmeticOperator.MULTIPLY.value,
             parameters=[
                 Formula(
-                    ArithmeticOperator.DIVIDE,
+                    ArithmeticOperator.DIVIDE.value,
                     [
                         Timeseries(
                             metric=Metric(
@@ -336,6 +322,136 @@ formula_tests = [
             groupby=[Column("transaction")],
         ),
         '((sum(foo){tag2:"tag_value2" AND (tag:"tag_value" OR tag:"tag_valueor")} / sum(bar)){tag3:"tag_value3"} * sum(pop)) by (transaction)',
+        id="test complex nested terms with OR condition",
+    ),
+    pytest.param(
+        Formula(
+            function_name="apdex",
+            parameters=[
+                Timeseries(
+                    metric=Metric(
+                        public_name="foo",
+                        entity="generic_metrics_distributions",
+                    ),
+                    aggregate="sum",
+                ),
+                500,
+            ],
+            filters=[Condition(Column("tag"), Op.EQ, "tag_value")],
+            groupby=[Column("transaction")],
+        ),
+        'apdex(sum(foo), 500){tag:"tag_value"} by (transaction)',
+        id="test arbitrary function with filters and groupby",
+    ),
+    pytest.param(
+        Formula(
+            function_name="apdex",
+            parameters=[
+                Formula(
+                    function_name=ArithmeticOperator.DIVIDE.value,
+                    parameters=[
+                        Timeseries(
+                            metric=Metric(
+                                public_name="foo",
+                                entity="generic_metrics_distributions",
+                            ),
+                            aggregate="sum",
+                        ),
+                        Timeseries(
+                            metric=Metric(
+                                public_name="bar",
+                                entity="generic_metrics_distributions",
+                            ),
+                            aggregate="sum",
+                        ),
+                    ],
+                ),
+                500,
+            ],
+            filters=[Condition(Column("tag"), Op.EQ, "tag_value")],
+            groupby=[Column("transaction")],
+        ),
+        'apdex((sum(foo) / sum(bar)), 500){tag:"tag_value"} by (transaction)',
+        id="test arbitrary function with filters, groupby, and terms",
+    ),
+    pytest.param(
+        Formula(
+            function_name="topK",
+            parameters=[
+                Timeseries(
+                    metric=Metric(
+                        public_name="foo",
+                        entity="generic_metrics_distributions",
+                    ),
+                    aggregate="sum",
+                ),
+                500,
+                "random",
+                "test",
+                4.2,
+            ],
+            filters=[Condition(Column("tag"), Op.EQ, "tag_value")],
+            groupby=[Column("transaction")],
+        ),
+        'topK(sum(foo), 500, random, test, 4.2){tag:"tag_value"} by (transaction)',
+        id="test arbitrary function with filters, groupby, and terms",
+    ),
+    pytest.param(
+        Formula(
+            function_name="multiply",
+            parameters=[
+                Formula(
+                    "apdex",
+                    [
+                        Timeseries(
+                            metric=Metric(
+                                public_name="foo",
+                                entity="generic_metrics_distributions",
+                            ),
+                            aggregate="sum",
+                        ),
+                        500,
+                    ],
+                ),
+                Formula(
+                    "apdex",
+                    [
+                        Timeseries(
+                            metric=Metric(
+                                public_name="foo",
+                                entity="generic_metrics_distributions",
+                            ),
+                            aggregate="sum",
+                        ),
+                        400,
+                    ],
+                ),
+            ],
+            filters=[Condition(Column("tag"), Op.EQ, "tag_value")],
+            groupby=[Column("transaction")],
+        ),
+        '(apdex(sum(foo), 500) * apdex(sum(foo), 400)){tag:"tag_value"} by (transaction)',
+        id="test arbitrary functions",
+    ),
+    pytest.param(
+        Formula(
+            "apdex",
+            [
+                Timeseries(
+                    metric=Metric(
+                        public_name="foo",
+                        entity="generic_metrics_distributions",
+                    ),
+                    aggregate="quantiles",
+                    aggregate_params=[0.5],
+                ),
+                500,
+            ],
+            filters=[Condition(Column("tag"), Op.EQ, "tag_value")],
+            groupby=[Column("transaction")],
+        ),
+        'apdex(quantiles(0.5)(foo), 500){tag:"tag_value"} by (transaction)',
+        id="test arbitrary functions",
     ),
 ]